[multiview] Update to v0.2.3
See release notes: https://github.com/swvn-dispatch/dispatcharr-multiview/releases/tag/v0.2.3
