+++ plugins/multiview/server.py
@@ -26,169 +26,29 @@ import sys
 
 from . import dispatcharr as _dispatcharr
 from . import layouts as _layouts
-from .ffmpeg import _parse_resolution, fps_string
+from .parameters import resolve_preset
 
 logger = logging.getLogger(__name__)
+
+_FPS_CHOICES = {"24", "25", "30", "50", "60", "30000/1001", "60000/1001"}
+
+
+def fps_string(settings: dict) -> str:
+    v = str(settings.get("output_fps") or "30")
+    return v if v in _FPS_CHOICES else "30"
+
+
+def _parse_resolution(settings: dict) -> tuple:
+    try:
+        w, h = (int(x) for x in (settings.get("output_resolution") or "1920x1080").split("x"))
+        return w, h
+    except Exception:
+        return 1920, 1080
 
 CHUNK_SIZE = 65536
 _WORKER = os.path.join(os.path.dirname(os.path.abspath(__file__)), "compositor_worker.py")
 
 _server_instance = None
-_mv_keepalives: dict = {}
-
-
-
-def _mv_keepalive_ensure(channel_id: str, proxy_server) -> "str | None":
-    """Register keepalive client (if absent), increment ref-count, and spawn drainer."""
-    import time as _t
-    import uuid as _u
-    if channel_id in _mv_keepalives:
-        cid, last, refs, drainer = _mv_keepalives[channel_id]
-        _mv_keepalives[channel_id] = (cid, last, refs + 1, drainer)
-        return cid
-    mgr = proxy_server.client_managers.get(channel_id)
-    if mgr is None:
-        return None
-    cid = str(_u.uuid4())
-    try:
-        if mgr.add_client(cid, "127.0.0.1", "multiview-keepalive", None, "mpegts", None):
-            drainer = None
-            try:
-                import gevent as _gv_ka
-                drainer = _gv_ka.spawn(_mv_keepalive_drain, channel_id, cid, proxy_server)
-            except Exception:
-                pass
-            _mv_keepalives[channel_id] = (cid, _t.time(), 1, drainer)
-            logger.info(f"Keepalive registered for {channel_id}")
-            return cid
-    except Exception as e:
-        logger.warning(f"Keepalive register failed for {channel_id}: {e}")
-    return None
-
-
-def _mv_keepalive_drain(channel_id: str, cid: str, proxy_server) -> None:
-    """Background greenlet: drain data for the keepalive client to prevent
-    Dispatcharr ghost-removal (last_active must stay fresh).
-
-    Only runs a StreamGenerator when ref_count == 0 (no active streaming
-    consumer).  While the streaming client is connected, it sleeps cheaply —
-    the streaming client's own generator keeps last_active current."""
-    try:
-        import gevent as _gv
-        from apps.proxy.live_proxy.output.ts.generator import StreamGenerator as _SG
-    except ImportError:
-        return
-
-    while True:
-        entry = _mv_keepalives.get(channel_id)
-        if entry is None or entry[0] != cid:
-            return
-
-        _, _, refs, _ = entry
-        if refs > 0:
-            # Streaming client is active; sleep, then refresh keepalive's
-            # last_active so Dispatcharr's ghost detector doesn't remove it.
-            _gv.sleep(1.0)
-            try:
-                import time as _t_ka
-                _mgr_ka = proxy_server.client_managers.get(channel_id)
-                if _mgr_ka and _mgr_ka.redis_client:
-                    _mgr_ka.redis_client.hset(
-                        f"live:channel:{channel_id}:clients:{cid}",
-                        "last_active", str(_t_ka.time()),
-                    )
-            except Exception:
-                pass
-            continue
-
-        # refs == 0: no streaming consumer — drain to keep last_active fresh.
-        buf = proxy_server.get_buffer(channel_id)
-        if buf is None:
-            _gv.sleep(0.5)
-            continue
-
-        mgr = proxy_server.client_managers.get(channel_id)
-        if mgr is None:
-            _gv.sleep(0.5)
-            continue
-
-        # Re-register if StreamGenerator._cleanup() removed the slot last iteration.
-        mgr.add_client(cid, "127.0.0.1", "multiview-keepalive", None, "mpegts", None)
-
-        try:
-            gen = _SG(
-                channel_id=channel_id,
-                client_id=cid,
-                client_ip="127.0.0.1",
-                client_user_agent="multiview-keepalive",
-                channel_initializing=False,
-                buffer=buf,
-            )
-            for _chunk in gen.generate():
-                _gv.sleep(0)
-                entry = _mv_keepalives.get(channel_id)
-                if entry is None or entry[0] != cid:
-                    return
-                if entry[2] > 0:
-                    # Streaming client reconnected; stop draining.
-                    break
-        except Exception:
-            _gv.sleep(0.1)
-
-
-def _mv_keepalive_release(channel_id: str) -> None:
-    """Decrement ref-count; TTL starts counting when it reaches zero."""
-    import time as _t
-    if channel_id in _mv_keepalives:
-        cid, _, refs, drainer = _mv_keepalives[channel_id]
-        _mv_keepalives[channel_id] = (cid, _t.time(), max(0, refs - 1), drainer)
-
-
-def _mv_keepalive_cleanup(proxy_server, ttl: float = 30.0) -> None:
-    """Remove keepalive clients idle (ref_count==0) longer than ttl seconds.
-    Kills drainer greenlet and removes Dispatcharr client slot."""
-    import time as _t
-    now = _t.time()
-    stale = [
-        u for u, (_, last, refs, _drainer) in list(_mv_keepalives.items())
-        if refs == 0 and now - last > ttl
-    ]
-    for uid in stale:
-        entry = _mv_keepalives.pop(uid, None)
-        if entry:
-            cid, _, _, drainer = entry
-            try:
-                if drainer is not None:
-                    drainer.kill(block=False)
-            except Exception:
-                pass
-            try:
-                mgr = proxy_server.client_managers.get(uid)
-                if mgr:
-                    mgr.remove_client(cid)
-                    logger.info(f"Keepalive expired and removed for {uid}")
-            except Exception:
-                pass
-
-
-def _mv_keepalive_shutdown(proxy_server) -> None:
-    """Kill all drainer greenlets and remove all keepalive clients immediately.
-    Called when the multiview composition stops to avoid channels lingering."""
-    entries = list(_mv_keepalives.items())
-    _mv_keepalives.clear()
-    for uid, (cid, _, _, drainer) in entries:
-        try:
-            if drainer is not None:
-                drainer.kill(block=False)
-        except Exception:
-            pass
-        try:
-            mgr = proxy_server.client_managers.get(uid)
-            if mgr:
-                mgr.remove_client(cid)
-                logger.info(f"Keepalive shutdown: removed {uid}")
-        except Exception:
-            pass
 
 
 def _python_exe() -> str:
@@ -233,11 +93,8 @@ def _lang_code(name: str) -> str:
     words = clean.split()
     if len(words) <= 1:
         return ((words[0] if words else "unk") + "   ")[:3].lower()
-    sig = [w for w in words if len(w) >= 2 and w.isupper()]
-    if sig:
-        return ("".join(sig) + "   ")[:3].lower()
-    initials = "".join(w[0] for w in words if w)
-    return (initials + "   ")[:3].lower()
+    parts = [w if (w.isupper() and len(w) >= 2) else w[0] for w in words]
+    return ("".join(parts) + "   ")[:3].lower()
 
 
 def _deduplicate_lang_codes(names: list) -> list:
@@ -254,23 +111,6 @@ def _deduplicate_lang_codes(names: list) -> list:
         else:
             result.append(code)
     return result
-
-
-def _audio_metadata_args(audio_source: str, channel_names: list, n: int) -> list:
-    args = []
-    if audio_source == "all":
-        lang_codes = _deduplicate_lang_codes(channel_names or [])
-        for i, (name, code) in enumerate(zip(channel_names or [], lang_codes)):
-            args += [f"-metadata:s:a:{i}", f"title={name}",
-                     f"-metadata:s:a:{i}", f"language={code}"]
-    else:
-        audio_idx = int(audio_source) if str(audio_source).isdigit() else 0
-        audio_idx = max(0, min(audio_idx, n - 1))
-        if channel_names and audio_idx < len(channel_names):
-            name = channel_names[audio_idx]
-            args += ["-metadata:s:a:0", f"title={name}",
-                     "-metadata:s:a:0", f"language={_lang_code(name)}"]
-    return args
 
 
 def _channel_logo(ch) -> "str | None":
@@ -388,14 +228,7 @@ class MultiviewServer:
             })
 
         encoder = settings.get("video_encoder") or "libx264"
-        _nvenc_presets = {"p1", "p2", "p3", "p4", "p5", "p6", "p7"}
-        _x264_presets  = {"ultrafast", "superfast", "veryfast", "faster", "fast", "medium", "slow"}
-        if encoder == "h264_nvenc":
-            saved = settings.get("encoder_preset")
-            preset = saved if saved in _nvenc_presets else "p4"
-        else:
-            saved = settings.get("encoder_preset")
-            preset = saved if saved in _x264_presets else "ultrafast"
+        preset = resolve_preset(encoder, settings.get("encoder_preset"))
         return {
             "out_w": out_w, "out_h": out_h, "fps": fps_string(settings),
             "bitrate": int(settings.get("output_bitrate") or 8000),
@@ -441,16 +274,28 @@ class MultiviewServer:
             start_response("404 Not Found", [("Content-Type", "text/plain")])
             return [b"Unknown channel\n"]
 
-        gen = _dispatcharr.live_stream(channel_uuid)
-        try:
-            first = next(gen)
-        except StopIteration:
-            # live_stream returned without yielding -- proxy not ready yet.
-            # Return 503 so the compositor worker sees a clean HTTP error and
-            # retries via its backoff, rather than "Invalid data found when
-            # processing input" from a 200 with an empty body.
-            start_response("503 Service Unavailable", [("Content-Type", "text/plain")])
-            return [b"Channel not ready\n"]
+        # Retry briefly while the live_proxy warms up the channel after container
+        # start. live_stream() cleans up its client registration in its finally
+        # block before raising StopIteration, so retrying is safe. Keep retries
+        # low (2) to avoid spamming logs when a channel is genuinely unavailable.
+        first = None
+        gen = None
+        for attempt in range(3):
+            gen = _dispatcharr.live_stream(channel_uuid)
+            try:
+                first = next(gen)
+                break
+            except StopIteration:
+                if attempt < 2:
+                    try:
+                        import gevent as _gv
+                        _gv.sleep(1.0)
+                    except ImportError:
+                        import time as _t
+                        _t.sleep(1.0)
+                else:
+                    start_response("503 Service Unavailable", [("Content-Type", "text/plain")])
+                    return [b"Channel not ready\n"]
 
         start_response("200 OK", [
             ("Content-Type", "video/mp2t"),
@@ -533,31 +378,36 @@ class MultiviewServer:
             sock.bind((self.host, self.port))
             sock.close()
         except OSError as e:
-            logger.error(f"Cannot bind to {self.host}:{self.port}: {e}")
+            logger.info(f"Multiview: port {self.port} already taken, skipping ({e})")
             return False
 
         try:
             from gevent import pywsgi
             import gevent as _gevent
-
-            def _run():
-                try:
-                    self._server = pywsgi.WSGIServer(
-                        (self.host, self.port), self.wsgi_app, log=None,
-                    )
-                    self.running = True
-                    set_server(self)
-                    self._server.serve_forever()
-                except Exception as e:  # noqa: BLE001
-                    logger.error(f"Multiview server crashed: {e}", exc_info=True)
-                finally:
-                    self.running = False
-
-            self._greenlet = _gevent.spawn(_run)
-            return True
         except ImportError:
             logger.error("gevent is not installed; cannot start multiview server")
             return False
+
+        def _run():
+            try:
+                self._server = pywsgi.WSGIServer(
+                    (self.host, self.port), self.wsgi_app, log=None,
+                )
+                self.running = True
+                set_server(self)
+                self._server.serve_forever()
+            except OSError as e:
+                # EADDRINUSE here means a concurrent worker won the race between
+                # our test-bind above and this re-bind -- expected on multi-worker
+                # startup, not an error.
+                logger.info(f"Multiview: port {self.port} taken by concurrent worker ({e})")
+            except Exception as e:  # noqa: BLE001
+                logger.error(f"Multiview server crashed: {e}", exc_info=True)
+            finally:
+                self.running = False
+
+        self._greenlet = _gevent.spawn(_run)
+        return True
 
     def stop(self):
         if self._server:
